+++ Vector_Math.h
@@ -4,13 +4,13 @@
 #include <vector>
 #include <memory>
 
-void vect_add(std::vector<float>& v1, std::vector<float>& v2, std::vector<float>& product);
-void resize(std::vector<float>& vect, float scale);
-void magnitude(std::vector<float>& object, float& mag);
-void displacement_vect(std::vector<float>& d, std::vector<float>& Acoords, int atom1, int atom2);
-void unit_vector(float& mag, std::vector<float> d, std::vector<float>& unitv);
-void dot(std::vector<float>& disp1, std::vector<float>& disp2, float& val);
-void cross(std::vector<float>& vect1, std::vector<float>& vect2, std::vector<float>& cprod);
+void vect_add(std::vector<double>& v1, std::vector<double>& v2, std::vector<double>& product);
+void resize(std::vector<double>& vect, double scale);
+void magnitude(std::vector<double>& object, double& mag);
+void displacement_vect(std::vector<double>& d, std::vector<double>& Acoords, int atom1, int atom2);
+void unit_vector(double& mag, std::vector<double> d, std::vector<double>& unitv);
+void dot(std::vector<double>& disp1, std::vector<double>& disp2, double& val);
+void cross(std::vector<double>& vect1, std::vector<double>& vect2, std::vector<double>& cprod);
